+++ plugins/codeamp/resourceconfig/errors.go
@@ -3,4 +3,5 @@ package resourceconfig
 const (
 	NilDependencyForExportErr = "required dependency for exporting missing %s"
 	FailedTypeAssertionErr    = "failed to assert type %s"
+	ObjectAlreadyExistsErr    = "%s object already exists"
 )
