+++ plugins/codeamp/resourceconfig/secret.go
@@ -20,14 +20,41 @@ type Secret struct {
 	Value string `yaml:"value"`
 }
 
-func CreateSecretConfig(config *string, db *gorm.DB, secret *model.Secret, project *model.Project, env *model.Environment) *SecretConfig {
+func CreateSecretConfig(db *gorm.DB, secret *model.Secret, project *model.Project, env *model.Environment) *SecretConfig {
 	return &SecretConfig{
-		db:                 db,
-		secret:             secret,
-		project:            project,
-		environment:        env,
-		BaseResourceConfig: BaseResourceConfig{config: config},
+		db:          db,
+		secret:      secret,
+		project:     project,
+		environment: env,
 	}
+}
+
+func (p *SecretConfig) Import(secret *Secret) error {
+	if p.db == nil || p.project == nil || p.environment == nil {
+		return fmt.Errorf(NilDependencyForExportErr, "db, project, environment")
+	}
+
+	// check if secret already exists
+	if err := p.db.Where("project_id = ? and environment_id = ? and key = ?", p.project.Model.ID, p.environment.Model.ID, secret.Key).Find(&model.Secret{}).Error; err == nil {
+		return fmt.Errorf(ObjectAlreadyExistsErr, "Secret")
+	}
+
+	newDBSecret := model.Secret{
+		Key: secret.Key,
+	}
+	if err := p.db.Create(&newDBSecret).Error; err != nil {
+		return err
+	}
+
+	newDBSecretValue := model.SecretValue{
+		Value:    secret.Value,
+		SecretID: newDBSecret.Model.ID,
+	}
+	if err := p.db.Create(&newDBSecretValue).Error; err != nil {
+		return err
+	}
+
+	return nil
 }
 
 func (s *SecretConfig) Export() (*Secret, error) {
